feat: add foreach loops, iterators, unpacking and tuple support
- add iterator protocol and foreach loop support
- implement list and tuple iterators
- add iterable-based destructuring assignment
- add call-site argument unpacking (*args)
- unify assignment logic into assign_target()
- support nested destructuring targets
- refactor assignment targets to use node-based dispatch
- replace parameter names with tokens for better diagnostics
- convert expression lists into tuples
- add tuple object implementation

+++ syntax/bs_objs/string/bs_string.h
@@ -27,7 +27,7 @@ public:
 
     [[nodiscard]] bs_obj_ptr subscript(interpreter &visitor, const bs_obj_ptr &index) const override;
 
-    [[nodiscard]] bs_obj_ptr slice(interpreter &visitor, int start, int end, int step) const override;
+    [[nodiscard]] bs_obj_ptr slice(interpreter &visitor, int64_t start, int64_t end, int64_t step) const override;
 
     [[nodiscard]] size_t len() const override;
